requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
line-bot-sdk==3.11.0